# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: JL Business Growth API
  version: 0.2.0
  description: HTTP contract for the JL团队生意成长管理系统.
servers:
  - url: /
...
paths:
...
  /api/team/members:
    get:
      operationId: listTeamMembers
      tags: [team-knowledge]
      summary: List the current user's team members
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createTeamMember
      tags: [team-knowledge]
      summary: Create a team member
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberRequest"
      responses:
        "201":
          description: Team member created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/team/members/{member_id}:
    get:
      operationId: getTeamMember
      tags: [team-knowledge]
      summary: Get a team member
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/MemberId"
      responses:
        "200":
          description: Team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: updateTeamMember
      tags: [team-knowledge]
      summary: Update a team member
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/MemberId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberRequest"
      responses:
        "200":
          description: Team member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMemberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteTeamMember
      tags: [team-knowledge]
      summary: Delete a team member
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/MemberId"
        - $ref: "#/components/parameters/PromoteChildren"
      responses:
        "204":
          description: Team member deleted
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/team/snapshots:
    get:
      operationId: listTeamSnapshots
      tags: [team-knowledge]
      summary: List team history snapshots
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Team snapshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSnapshotListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createTeamSnapshot
      tags: [team-knowledge]
      summary: Capture a manual team snapshot
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamSnapshotRequest"
      responses:
        "201":
          description: Team snapshot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSnapshotResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/team/snapshots/{snapshot_id}:
    get:
      operationId: getTeamSnapshot
      tags: [team-knowledge]
      summary: Get a read-only team snapshot
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: Team snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSnapshotResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/knowledge:
    get:
      operationId: listKnowledgeItems
      tags: [team-knowledge]
      summary: List knowledge items
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Knowledge items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItemListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createKnowledgeItem
      tags: [team-knowledge]
      summary: Create a knowledge item
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeItemRequest"
      responses:
        "201":
          description: Knowledge item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItemResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/knowledge/{knowledge_id}:
    get:
      operationId: getKnowledgeItem
      tags: [team-knowledge]
      summary: Get a knowledge item
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/KnowledgeId"
      responses:
        "200":
          description: Knowledge item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItemResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      operationId: updateKnowledgeItem
      tags: [team-knowledge]
      summary: Update a knowledge item
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/KnowledgeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeItemRequest"
      responses:
        "200":
          description: Knowledge item updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItemResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteKnowledgeItem
      tags: [team-knowledge]
      summary: Delete a knowledge item
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/KnowledgeId"
      responses:
        "204":
          description: Knowledge item deleted
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/learning-sessions:
    get:
      operationId: listLearningSessions
      tags: [team-knowledge]
      summary: List learning sessions
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Learning sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSessionListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: createLearningSession
      tags: [team-knowledge]
      summary: Record an item or unallocated learning session
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LearningSessionRequest"
      responses:
        "201":
          description: Learning session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LearningSessionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/learning-sessions/{session_id}:
    delete:
      operationId: deleteLearningSession
      tags: [team-knowledge]
      summary: Delete a learning session
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          description: Learning session deleted
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/files:
    get:
      operationId: listFiles
      tags: [team-knowledge]
      summary: List uploaded files
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Uploaded files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      operationId: uploadFile
      tags: [team-knowledge]
      summary: Upload an allowed file without parsing its contents
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                category:
                  type: string
                  enum: [DREAM_IMAGE, KNOWLEDGE_DOCUMENT, KNOWLEDGE_IMAGE]
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/files/{file_id}:
    delete:
      operationId: deleteFile
      tags: [team-knowledge]
      summary: Delete an uploaded file
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/FileId"
      responses:
        "204":
          description: File deleted
        default:
          $ref: "#/components/responses/ErrorResponse"
  /api/files/{file_id}/content:
    get:
      operationId: getFileContent
      tags: [team-knowledge]
      summary: Preview or download an uploaded file
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/FileId"
        - name: disposition
          in: query
          schema:
            type: string
            enum: [inline, attachment]
            default: inline
      responses:
        "200":
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bos_session
  parameters:
...
    DateFrom:
      name: from
      in: query
      schema:
        type: string
        format: date
    DateTo:
      name: to
      in: query
      schema:
        type: string
        format: date
...
    MemberId:
      name: member_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SnapshotId:
      name: snapshot_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    KnowledgeId:
      name: knowledge_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FileId:
      name: file_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PromoteChildren:
      name: promote_children
      in: query
      schema:
        type: boolean
        default: false
...
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        request_id:
          type: string
...
    TeamMember:
      type: object
      required: [id, member_code, name, status, sort_order, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        member_code: { type: string, minLength: 1, maxLength: 100 }
        parent_id: { type: string, format: uuid, nullable: true }
        name: { type: string }
        joined_on: { type: string, format: date, nullable: true }
        rank: { type: string, nullable: true }
        city: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, INACTIVE] }
        note: { type: string, nullable: true }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TeamMemberRequest:
      type: object
      required: [name]
      properties:
        member_code: { type: string, minLength: 1, maxLength: 100 }
        parent_id: { type: string, format: uuid, nullable: true }
        name: { type: string, minLength: 1, maxLength: 200 }
        joined_on: { type: string, format: date, nullable: true }
        rank: { type: string, nullable: true }
        city: { type: string, nullable: true }
        status: { type: string, enum: [ACTIVE, INACTIVE], default: ACTIVE }
        note: { type: string, nullable: true }
        sort_order: { type: integer, default: 0 }
    TeamMemberResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/TeamMember" }
        request_id: { type: string }
    TeamMemberListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/TeamMember" }
        request_id: { type: string }
...
    TeamSnapshotRequest:
      type: object
      properties:
        snapshot_month: { type: string, format: date }
        snapshot_type: { type: string, enum: [AUTO, MANUAL], default: MANUAL }
        captured_late: { type: boolean, default: false }
    TeamSnapshotResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/TeamSnapshot" }
        request_id: { type: string }
    TeamSnapshotListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/TeamSnapshot" }
        request_id: { type: string }
...
    KnowledgeItemRequest:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1, maxLength: 300 }
        type: { type: string, enum: [AUDIO, VIDEO, BOOK, EVENT, MEETING, PHP, MENTOR, PRODUCT, OTHER], default: OTHER }
        raw_text: { type: string, nullable: true }
        summary: { type: string, nullable: true }
        understanding: { type: string, nullable: true }
        action_items: { type: string, nullable: true }
        source_url: { type: string, nullable: true }
        learned_on: { type: string, format: date, nullable: true }
        status: { type: string, enum: [NOT_STARTED, IN_PROGRESS, COMPLETED], default: NOT_STARTED }
        progress_current: { type: number, nullable: true }
        progress_total: { type: number, nullable: true }
        progress_unit: { type: string, nullable: true }
        tags: { type: array, items: { type: string } }
        file_ids: { type: array, items: { type: string, format: uuid } }
    KnowledgeItemResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/KnowledgeItem" }
        request_id: { type: string }
    KnowledgeItemListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/KnowledgeItem" }
        request_id: { type: string }
...
    LearningSessionRequest:
      type: object
      required: [activity_type, activity_date, minutes]
      properties:
        knowledge_item_id: { type: string, format: uuid, nullable: true }
        activity_type: { type: string, enum: [READING, AUDIO] }
        activity_date: { type: string, format: date }
        minutes: { type: integer, minimum: 0 }
        source: { type: string, enum: [ITEM, DAILY_UNALLOCATED], default: ITEM }
        note: { type: string, nullable: true }
    LearningSessionResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/LearningSession" }
        request_id: { type: string }
    LearningSessionListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/LearningSession" }
        request_id: { type: string }
...
    FileResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/FileAsset" }
        request_id: { type: string }
    FileListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/FileAsset" }
        request_id: { type: string }
...